refactor(server): trim connector barrel values
