+++ apps/server/src/connector/index.ts
@@ -4,9 +4,6 @@ import {
   type ConnectorQuestionBridgeHandler,
 } from "./process-driver.js";
 
-export { ServerConnectorDefinitions } from "./definitions.js";
-export { ServerConnectorDiscovery } from "./discovery.js";
-export { ServerConnectorRegistry } from "./registry.js";
 export type { ConnectorQuestionBridgeHandler };
 
 export interface ConnectorEndpointDriverOptions {
