+++ apps/server/test/connector/registry-smoke-verify.test.ts
@@ -9,11 +9,9 @@ import {
   SqliteStorageAdapter,
   Storage,
 } from "@openomni/session";
-import {
-  ServerConnectorDiscovery,
-  ServerConnectorRegistry,
-  ServerConnectorDefinitions,
-} from "../../src/connector/index.js";
+import { ServerConnectorDefinitions } from "../../src/connector/definitions.js";
+import { ServerConnectorDiscovery } from "../../src/connector/discovery.js";
+import { ServerConnectorRegistry } from "../../src/connector/registry.js";
 
 async function availableCodexCandidate(): Promise<AppConnector.Definition> {
   const connector = ServerConnectorDefinitions.get("app.codex");
